Implemented use cases to move business logic out of HomeScreenUiState;
