+++ app/src/main/java/com/coryrdunn/pscodeexercise/domain/model/DriverWithShipment.kt
@@ -1,0 +1,6 @@
+package com.coryrdunn.pscodeexercise.domain.model
+
+data class DriverWithShipment (
+    val driver: String = "",
+    val shipment: String = ""
+)
